Auth controller bug fix 1.1

+++ src/main/java/com/adminmodule/controller/CustomerController.java
@@ -2,6 +2,7 @@ package com.adminmodule.controller;
 
 import com.adminmodule.domain.Customer;
 import com.adminmodule.security.AccountDetailsService;
+import com.adminmodule.security.JWTUtility;
 import com.adminmodule.service.CustomerService;
 import com.adminmodule.service.dto.AddressDTO;
 import com.adminmodule.service.dto.CustomerDTO;
@@ -17,23 +18,31 @@ import java.util.List;
 @RequestMapping("api/v1/customer")
 public class CustomerController {
 
+    private final JWTUtility jwtUtility;
+
     private final CustomerService customerService;
 
     @Autowired
-    public CustomerController(CustomerService customerService) {
+    public CustomerController(CustomerService customerService, JWTUtility jwtUtility) {
         this.customerService = customerService;
+        this.jwtUtility = jwtUtility;
     }
     @GetMapping("/all")
     public ResponseEntity<?> getAllCustomers(){
-
         List<Customer> customers = customerService.getAllCustomers();
         return new ResponseEntity<>(customers, HttpStatus.OK);
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<?> getCustomerById(@PathVariable("id") Long id){
-        CustomerDTO customerDTO = customerService.getCustomerById(id);
-        return new ResponseEntity<>(customerDTO, HttpStatus.OK);
+    public ResponseEntity<?> getCustomerById(@PathVariable("id") Long id, @RequestHeader("Authorization") String authorizationHeader){
+        String jwtToken = authorizationHeader.substring(7);
+        String username = jwtUtility.getUsernameFromToken(jwtToken);
+        CustomerDTO customerDTO = customerService.getCustomerByUsername(username);
+        CustomerDTO customer = customerService.getCustomerById(id);
+        if (customer.getId()==customerDTO.getId()){
+            return new ResponseEntity<>(customer, HttpStatus.OK);
+        }
+        return new ResponseEntity<>(HttpStatus.FORBIDDEN);
     }
 
     @PostMapping()
@@ -43,26 +52,52 @@ public class CustomerController {
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<?> updateCustomer(@PathVariable("id") Long id,@RequestBody CustomerDTO customerDTO){
-       CustomerDTO newCustomer =  customerService.updateCustomer(customerDTO, id);
+    public ResponseEntity<?> updateCustomer(@PathVariable("id") Long id,@RequestBody CustomerDTO customerDTO, @RequestHeader("Authorization") String authorizationHeader){
+        String jwtToken = authorizationHeader.substring(7);
+        String username = jwtUtility.getUsernameFromToken(jwtToken);
+        CustomerDTO customerDTO1 = customerService.getCustomerByUsername(username);
+        if (customerDTO1.getId()!=id){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
+        CustomerDTO newCustomer =  customerService.updateCustomer(customerDTO, id);
         return new ResponseEntity<>(newCustomer ,HttpStatus.OK);
     }
 
     @DeleteMapping("/{id}")
-    public ResponseEntity<?> deleteCustomer(@PathVariable("id") Long id){
+    public ResponseEntity<?> deleteCustomer(@PathVariable("id") Long id, @RequestHeader("Authorization") String authorizationHeader){
+        String jwtToken = authorizationHeader.substring(7);
+        String username = jwtUtility.getUsernameFromToken(jwtToken);
+        CustomerDTO customerDTO = customerService.getCustomerByUsername(username);
+        if (customerDTO.getId()!=id){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         customerService.deleteCustomer(id);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
     @PutMapping("/shipping-address/{id}")
     public ResponseEntity<?> updateShippingAddress(@PathVariable("id") Long id,
-                                                   @RequestBody AddressDTO addressDTO){
+                                                   @RequestBody AddressDTO addressDTO,
+                                                   @RequestHeader("Authorization") String authorizationHeader){
+        String jwtToken = authorizationHeader.substring(7);
+        String username = jwtUtility.getUsernameFromToken(jwtToken);
+        CustomerDTO customerDTO = customerService.getCustomerByUsername(username);
+        if (customerDTO.getId()!=id){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         return ResponseEntity.ok(customerService.updateShippingAddress(addressDTO, id));
     }
 
     @PutMapping("/billing-address/{id}")
     public ResponseEntity<?> updateBillingAddress(@PathVariable("id") Long id,
-                                                   @RequestBody AddressDTO addressDTO){
+                                                   @RequestBody AddressDTO addressDTO,
+                                                  @RequestHeader("Authorization") String authorizationHeader){
+        String jwtToken = authorizationHeader.substring(7);
+        String username = jwtUtility.getUsernameFromToken(jwtToken);
+        CustomerDTO customerDTO = customerService.getCustomerByUsername(username);
+        if (customerDTO.getId()!=id){
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
+        }
         return ResponseEntity.ok(customerService.updateBillingAddress(addressDTO, id));
     }
 }
